Use the "version 2" child theme, too.
Note that the `slug` property is needed to match the `composer.json`
file from https://github.com/glocalcoop/activist-network-composer since
its inline `package` declarations use that name, even though the GitHub
repository for the theme have a different name. The `slug` property in
the blueprint file is non-standard but is correctly parsed by the PHP
helper script `vvv-init.php` in this repository.

The easier way to deal with this is probably to consolidate the theme
repositories themselves since a "version 2" should just be handled by
git, but ah well.